mcp-servers/python/qr_code_server/pyproject.toml
[project]
name = "qr-code-server"
version = "0.1.0"
description = "MCP server that provides QR code generation and decoding capabilities"
authors = [
  { name = "Matheus Cafalchio", email = "mcafalchio@gmail.com" }
]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.14.4",
    "mcp>=1.26.0",
    "numpy>=2.4.2",
    "opencv-python-headless>=4.13.0.90",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "qrcode[pil]>=8.2",
]

[project.optional-dependencies]
dev = [
    "black>=26.1.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qr_code_server"]

[project.scripts]
qr-code-server = "qr_code_server.server:main"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "W", "F", "B", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=qr_code_server --cov-report=term-missing"
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "black>=25.11.0",
    "ipython>=9.7.0",
    "puremagic>=1.30",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.6",
]
